+++ v3/src/geom/point/Divide.js
@@ -1,0 +1,9 @@
+var Divide = function (point, x, y)
+{
+    point.x /= x;
+    point.y /= y;
+
+    return point;
+};
+
+module.exports = Divide;