+++ v3/src/geom/point/Multiply.js
@@ -1,0 +1,9 @@
+var Multiply = function (point, x, y)
+{
+    point.x *= x;
+    point.y *= y;
+
+    return point;
+};
+
+module.exports = Multiply;